added unlock lesson to appropriate view
