+++ openedx/features/genplus_features/genplus_learning/api/v1/urls.py
@@ -17,7 +17,6 @@ lesson_viewset_router.register('lessons', ProgramViewSet, basename='lessons')
 app_name = 'genplus_learning_api_v1'
 
 urlpatterns = [
-    url(r'^lessons/unlock/(?P<pk>\d+)/$', ProgramViewSet.as_view({"put": "unlock_lesson"})),
     url(r'^class-students/(?P<group_id>\w+)/$', ClassStudentViewSet.as_view({'get': 'list'})),
     path('', include(lesson_viewset_router.urls)),
 ]
